Keep the Gravito-Orbit physics step in double precision
The lab's traits declared kPhysicsStep as float while the model's
kLorentzPhysicsStep and StepLorentzParticle contract are double. MSVC
flags the narrowing under /WX (CI builds with warnings as errors; the
local presets do not), and the truncation would also have changed the
step value the model receives. The shell's traits contract now states
that kPhysicsStep may be float or double. Verified with a local
TINY2D_WARNINGS_AS_ERRORS=ON build over every target.

+++ Sandbox/lorentz_particle_simulation.cc
@@ -470,7 +470,8 @@ struct LorentzLabTraits {
   using Config = tiny2d::sandbox::LorentzParticleConfig;
   using State = tiny2d::sandbox::LorentzParticleState;
   using Status = tiny2d::sandbox::LorentzParticleStatus;
-  static constexpr float kPhysicsStep = tiny2d::sandbox::kLorentzPhysicsStep;
+  // Double, matching the model's double-precision step contract.
+  static constexpr double kPhysicsStep = tiny2d::sandbox::kLorentzPhysicsStep;
   static constexpr double kMaximumFrameTime = 0.25;
   static constexpr const char* kInvalidHistoryTimeMessage =
       "The particle model produced an invalid history time.";
